Engine: updates to support Changesets and Changeset History
This commit contains the changes to enable Changesets
(e.g. CRUD, promotion..) and Changeset History UI components
to work within the engine.  Both of these pages require
the user to be within the context of an organization; therefore,
if they are not, the page will render a message indicating to
select an org.